add a new structure
