+++ test/fixture/client.js
@@ -1,0 +1,6 @@
+const infoUser = {
+  subscriptionId: 123456,
+  apiToken: 'ar425598573462y24ec1ceee728981663',
+};
+
+export default infoUser;
