+++ test/fixture/entries/dataEntries.js
@@ -4,9 +4,5 @@ const dataEntryCreate = {
   notes: 'Issue',
   taskId: 14356973,
 };
-const infoUser = {
-  subscriptionId: 123456,
-  apiToken: 'ar425598573462y24ec1ceee728981663',
-};
 
-export { dataEntryCreate, infoUser };
+export default dataEntryCreate;
